requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
pydub==0.25.1
youtube-transcript-api==0.6.2
yt-dlp==2023.10.13
google-generativeai==0.3.2
assemblyai==0.21.0
pydantic==2.5.0